+++ src/main/java/com/auth0/json/mgmt/actions/ActionsPageDeserializer.java
@@ -4,6 +4,9 @@ import com.auth0.json.mgmt.PageDeserializer;
 
 import java.util.List;
 
+/**
+ * Parses a paged response into its {@linkplain ActionsPage} representation.
+ */
 public class ActionsPageDeserializer extends PageDeserializer<ActionsPage, Action> {
 
     protected ActionsPageDeserializer() {
