--- src/main/java/com/auth0/json/mgmt/actions/BindingReference.java
@@ -1,31 +1,0 @@
-package com.auth0.json.mgmt.actions;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class BindingReference {
-
-    @JsonProperty("type")
-    private String type;
-    @JsonProperty("value")
-    private String value;
-
-    @JsonCreator
-    public BindingReference(@JsonProperty("type") String type,
-                            @JsonProperty("value") String value) {
-        this.type = type;
-        this.value = value;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
