+++ src/main/java/com/auth0/json/mgmt/actions/BindingsPage.java
@@ -7,6 +7,11 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import java.util.List;
 
+/**
+ * Represents a page of an action trigger binding.
+ * @see Binding
+ * @see com.auth0.client.mgmt.ActionsEntity
+ */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(using = BindingsPageDeserializer.class)
