+++ src/main/java/com/auth0/json/mgmt/actions/ServiceStatus.java
@@ -4,6 +4,10 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+/**
+ * Represents the status of the actions service.
+ * @see com.auth0.client.mgmt.ActionsEntity
+ */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ServiceStatus {
@@ -11,6 +15,9 @@ public class ServiceStatus {
     @JsonProperty("status")
     private String status;
 
+    /**
+     * @return the status of the actions service.
+     */
     public String getStatus() {
         return status;
     }
