+++ src/main/java/com/auth0/json/mgmt/actions/VersionsPageDeserializer.java
@@ -4,6 +4,9 @@ import com.auth0.json.mgmt.PageDeserializer;
 
 import java.util.List;
 
+/**
+ * Parses a paged response into its {@linkplain VersionsPage} representation.
+ */
 public class VersionsPageDeserializer extends PageDeserializer<VersionsPage, Version> {
 
     protected VersionsPageDeserializer() {
